requirements.txt
# Sentinel-2 NDVI Analysis - Python Dependencies
# Install with: pip install -r requirements.txt

# Core geospatial libraries
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Cloud storage & data access
boto3>=1.28.0
pystac-client>=0.7.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Note: GDAL (for build_vrt.py)
# Note: GDAL should be installed system-wide or via conda
# conda install -c conda-forge gdal
# If using pip, try: pip install gdal (may require system GDAL)

# Jupyter (for running notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0
